Fix for Jira Issue NMS-6959
test failure: org.opennms.netmgt.collectd.DuplicatePrimaryAddressTest

In order to test if Collectd is able to handle data collection after
adding a new node with a duplicate primary IP, the test checks if the
new node is scheduled for data collection, and the collections happen at
least 5 times during the 6 seconds time window. On a slow machine, it
could be possible that the amount of expected collections is different,
so I've increased the time window to be 10 seconds, expecting that at
least 5 collections have been completed for the new node.