Improving the format for service-configuration.xml
Now, there is no need to comment a service definition when you want to
disable a service.

To disable a service, just add enabled="false" as an attribute to the
service tag and that's it, for example:

<service enabled="false">
    <name>OpenNMS:Name=Capsd</name>
    <class-name>org.opennms.netmgt.capsd.jmx.Capsd</class-name>
    <invoke at="start" pass="0" method="init"/>
    <invoke at="start" pass="1" method="start"/>
    <invoke at="status" pass="0" method="status"/>
    <invoke at="stop" pass="0" method="stop"/>
</service>

The default value is 'true'. If the attribute is not specified, it is
assumed to be 'true'.

The current content of the file has been updated to use this new
attribute for the services that are going to be disabled by default.