Fix for Bug NMS-7243
XML collector in JSON mode assumes all element content is String.

The AbstractJsonCollectionHandler was asuming that the content is always
a String which is not correct. Also, it was assuming it never will be
null, which is not correct either, so I fixed the code and I've added a
JUnit test.