Fix for Bug NMS-7432
Normalize the HTTP Host Header with the new HttpClientWrapper.

a) A request interceptor has been added to the HttpClientWrapper to
suppress the default port from the Host header if exist.
b) The wrapper has an option to use LAX Redirect
c) The PSM adds an interceptor for removing the host header if the
virtual-host attribute is not specified on the page sequence.