Fix for Bug NMS-2995 (Part II)
Improving the solution to make the JUnit tests happy.

For the same reason why I changed the code to make SNMPv3 Informs works,
the tests were failling. Basically, cloning a PDU is a bad thing, so we
should "reuse" the original PDU for preparing the response and then call
"trapReceived". Of course, preparing the response means changing some
PDU params, that must be "restored" before calling "trapReceived".

The solution was good for the JUnit tests, and it also was verified on a
live environment with real SNMPv2/SNMPv3 informs.