Fix for Bug NMS-7650
XML data collection with HTTP POST requests is not working.

The method parseRequest from the AbstractXmlCollectionHandler class was
not passing the method (i.e. GET or POST, and it defaults to GET if it
is configured) to the formatted version of the Request object, which was
preventing sending an actual POST to the target device.

A JUnit test has been added to verify the proper behavior of the POST
request against a test web-app.