Adding an optional element called v1-agent-address to the snmp-trap-sink.
If the parameter is not supplied, the IP address from the incoming alarm
will be used as the agent address attribute when sending an SNMPv1 trap.

If specified, it expects to be a SPEL expression that should return a
string with the IP address to use. If the SPEL parser can't parse the
expression, the the IP address from the incoming alarm will be used.