Improving SnmpPeerFactory to avoid version conflicts on SnmpTrapConfig.
The scope of the following method was changed from private to public:
org.opennms.netmgt.config.SnmpPeerFactory.getAgentConfig(InetAddress, int)

With the above method, you can retrieve the SNMP configuration for a
given IP address and a specific SNMP version, which is exactly what I
need.

Thanks to this, now all the tests are passing, and the SNMP version will
always be the same for the SnmpTrapConfig and the related
SnmpAgentConfig.