HZN-584: Create WS-Management documentation for OpenNMS Administrator
Build structure for performance data collection. The structure follows the same pattern as in Service Assurance. The
content is migrated from the Google Docs draft file to Asciidoc format. The description for the WS-Man detector is
separated in a dedicated file. There is a good chance this section will move to Provisioning detectors.

Added helpful commands for troubleshooting WinRM connectivity between Linux and Microsoft Windows Server. Added infos
and hints running with Microsoft Server 2012.