Fix for Bug NMS-8194 - Part II
For PUT/DELETE (and POST where 202 doesn't apply), we decided to return
a 204 (No Content), because there is no data returned to the client on
those cases.

Also, the GET requests have been improved to return NOT_FOUND (when the
final resource doesn't exist), or BAD_REQUEST (when an intermediate
resource doesn't exist, like the node for a given IP).

All the JUnit tests have been updated, and they are passing.