Fix for Enhancement NMS-8336
Set the required fields when not specified while adding events through ReST
Now, you don't need to set the time when pusing an event through ReST,
considering the restrictions on the date format. This is actually
consistent with how the EventBuilder works, which is widely used on
several parts of the code.