Fix for Bug NMS-7854
Translated events with a new UEI do not receive the description,
severity and alarm-data of the new event if new UEI's eventconf is
present after original UEI's eventconf.

The workaround of reordering the event definitions work for non-trap
events. For traps events, there are some unwanted results that are
affecting other components like the Passive Status Keeper.

For this reason, several fields (specially snmp) have to be set to null
when cloning the original event to avoid surprises.