NMS-8499 filter out JMX options on install
Normally, bootstrap attempts to load the spring-security JAAS context
once it is available at startup.  However, during install, it never
initializes because it's not a "real" OpenNMS launch, so no context
is available.

If you have pre-configured opennms.conf to modify the default jmxremote
settings, these will be passed to the installer as well, so it attempts
to initialize the JMX context but fails and dies with an error about
Error: Password file not found: $JAVA_HOME/jre/lib/management/jmxremote.password

So, the easiest fix is to just not pass those options to the installer.
:)