Post NMS-8646 Fix
chooseresource.htm doesn't exist anymore, so the netutils topology
plugin required a minor fix. I've updated chooseresource.jsp to support
the same parameters to avoid big changes on that plugin.

There is another class called DefaultResourceListPresenter.java from
features/gwt-graph-resource-list that also has a reference to the old
page, but that is part of a component that will be removed eventually.