Fix for Bug NMS-7106
If a target node is rebooted the RRD/JRB files contains spikes because
the sysUpTime check is not working.

Note: the sysUpTime check was working, and that was preventing the
persistence of the current collection-set if a restart was detected.
This is crucial for avoding spikes. Now, the counters cache was not
reinitialized which is the reason of this change.

The fix was tested on my lab and it works as expected.