HZN-1092: Upgrade to Karaf 4.1.2 (#1698)
* HZN-1092: Upgrade to Karaf 4.1.1

Spring 4.2 and earlier have been moved into a 'spring-legacy' feature
repository so include that repo in every project that needs it.

* Created a paxWebVersion build property

* NMS-9363: Upgrade to osgi-pluginmanager compatible with Pax Web 4 and 6

* HZN-1092: Downgrade Pax Web to a version that works with our HTTP bridge

* HZN-1092: Removed repo refresh in KarafExtender

Removed manual repo refresh from KarafExtender to avoid
refreshing a repo that already has installed features which is now
illegal in Karaf 4.1.  Bumped Pax Exam to latest version 4.11.0.
Other minor fixes.

https://issues.apache.org/jira/browse/KARAF-4060
https://issues.apache.org/jira/browse/KARAF-5223
https://issues.apache.org/jira/browse/KARAF-5251

* HZN-1092: Fixed console output parsing

Karaf 4.1 has jline ANSI escape codes on some output so we need
to change some test assertions.

* HZN-1092: Updated to Karaf 4.1.2, added workaround for KARAF-5384

The workaround involves adding our Apache MINA bundle to the startup
bundle list for Karaf in startup.properties. This has to be done in
the container in container/standalone and in the Pax Exam tests so
that the bundle is available before the container starts in Pax Exam.

https://issues.apache.org/jira/browse/KARAF-5384

* HZN-1092: Updated Minion container to match Karaf distro

The Minion karaf-maven-plugin configuration was missing many
configuration options, making the /lib, /lib/endorsed, and /system
directories incomplete. This has been fixed and the MINA JAR was
added to the /system repo to work around KARAF-5384.

https://issues.apache.org/jira/browse/KARAF-5384

* Clean up JRadius dependency POM

* HZN-1092: Disable Bouncy Castle security in Apache MINA SSHD

Karaf's SSHD was failing because it was trying to classload
Bouncy Castle out of the system classpath so disable the
Bouncy Castle provider.

* HZN-1092: Shade/relocate bouncycastle inside jradius-extended

Created a shaded JAR project to combine jradius-extended and
a relocated bouncycastle to avoid bad side-effects of having
bouncycastle in the classpath (like NMS-7959).