NMS-9816: Add Disk IO monitoring capabilities with Net-SNMP
Enable disk IO metrics for read/writes in bytes and operations in default system definition to get collected by default.
Added metrics for disk load in percent for 1, 5 and 15-minute averages and created RRD graph definitions for each value
as percentage graph and an aggregated graph with all three values similar to a load average graph. Improved the graph
definitions for existing diskio read writes with areas and lines.

Added 64-bit counters for reading and write bytes and if they are supported, suppress the graph with the 32-bit counter.