NMS-10296: make sure a single node matches
Because of `opennms.minion.provisioning.foreignSourcePattern` it is
difficult to limit the query to only matching foreign sources. This
patch changes it to pull all nodes that match the appropriat foreign
IDs, and then indexes them by a combination of foreignId and
location, since that should be unique regardless of how the
foreignSourcePattern is set.